v1.5 - #16 cleaner debug log when ' is used
Upstream TagUI project has a limitation in live mode. It was a tradeoff to enable dynamic variables working for selectors and parameters in live mode. As a result, TagUI for Python sends the following string to TagUI when a single quote ' is used as parameter (non-identifier parameter).

```
'+"\'"+'
```

With the issue https://github.com/kelaberetiv/TagUI/issues/465 raised by user, an improvement is made upstream, using a solution for a similar problem while working on this personal side project. Thus a commit can now be made here that replaces single quote ' for non-identifier parameter with

```
\'
```

This may seem like a small improvement, but it helps clarify in debug log tagui_python.log when t.debug(True) is set. Otherwise, whenever there is a ' it would result in some roundabout escape sequence above due to a limitation in upstream live mode. This commit fixes that by escaping ' to \' following standard convention.