#373 - register hide command window idea
committing a great idea from @jesuisme on hiding command prompt windows on Windows.

However the idea seems to only work on Windows 11. Since it can't be replicated on Windows 10, and have issues for Mac/Linux, giving the code change a pass.

However, still good to register this idea and Larry's contribution. 1/2 commits to commit.

Co-Authored-By: Larry Davis <3903655+jesuisme@users.noreply.github.com>